consolidate all activity event elems into shared directory. better apollo projectEvent caching
