+++ src/components/activityEventElems/v1/index.ts
@@ -1,0 +1,3 @@
+export * from './ReservesEventElem'
+export * from './TapEventElem'
+export * from './V1ConfigureEventElem'
