+++ src/components/activityEventElems/v2v3/DistributeReservedTokensElem.tsx
@@ -1,5 +1,4 @@
 import { Trans } from '@lingui/macro'
-import { ActivityEvent } from 'components/activityEventElems/ActivityElement'
 import EthereumAddress from 'components/EthereumAddress'
 import { V1ProjectContext } from 'contexts/v1/Project/V1ProjectContext'
 import { ProjectEventsQuery } from 'generated/graphql'
@@ -8,27 +7,37 @@ import { useContext } from 'react'
 import { formatWad, fromWad } from 'utils/format/formatNumber'
 import { tokenSymbolText } from 'utils/tokenSymbolText'
 
-export default function ReservesEventElem({
+import { ActivityEvent } from '../ActivityElement'
+
+export default function DistributeReservedTokensEventElem({
   event,
 }: {
-  event: ProjectEventsQuery['projectEvents'][0]['printReservesEvent']
+  event: ProjectEventsQuery['projectEvents'][0]['distributeReservedTokensEvent']
 }) {
   const { tokenSymbol } = useContext(V1ProjectContext)
 
-  // Load individual DistributeToTicketMod events, emitted by internal transactions of the PrintReserves transaction
+  // Load individual DistributeToReservedTokenSplit events, emitted by internal transactions of the DistributeReservedTokens transaction
   const { data: distributeEvents } = useSubgraphQuery(
-    event?.id
-      ? {
-          entity: 'distributeToTicketModEvent',
-          keys: ['id', 'timestamp', 'txHash', 'modBeneficiary', 'modCut'],
-          orderDirection: 'desc',
-          orderBy: 'modCut',
-          where: {
-            key: 'printReservesEvent',
+    {
+      entity: 'distributeToReservedTokenSplitEvent',
+      keys: [
+        'id',
+        'timestamp',
+        'txHash',
+        'beneficiary',
+        'tokenCount',
+        'projectId',
+      ],
+      orderDirection: 'desc',
+      orderBy: 'tokenCount',
+      where: event?.id
+        ? {
+            key: 'distributeReservedTokensEvent',
             value: event.id,
-          },
-        }
-      : null,
+          }
+        : undefined,
+    },
+    {},
   )
 
   if (!event) return null
@@ -47,36 +56,29 @@ export default function ReservesEventElem({
         </Trans>
       }
       subject={
-        <div className="text-base font-medium text-grey-900 dark:text-slate-100">
-          {formatWad(event.count, { precision: 0 })}{' '}
+        <div className="text-base font-medium">
+          {formatWad(event.tokenCount, { precision: 0 })}{' '}
           {tokenSymbolText({
             tokenSymbol,
             capitalize: false,
-            plural: parseInt(fromWad(event.count) || '0') !== 1,
+            plural: parseInt(fromWad(event.tokenCount) || '0') !== 1,
           })}
         </div>
       }
       extra={
         <div>
           {distributeEvents?.map(e => (
-            <div
-              key={e.id}
-              style={{
-                display: 'flex',
-                justifyContent: 'space-between',
-                alignItems: 'baseline',
-              }}
-              className="text-sm"
-            >
+            <div key={e.id} className="flex items-baseline justify-between">
               <div>
                 <EthereumAddress
                   className="text-grey-900 dark:text-slate-100"
-                  address={e.modBeneficiary}
+                  address={e.beneficiary}
                 />
                 :
               </div>
-              <div className="text-grey-500 dark:text-grey-300">
-                {formatWad(e.modCut, { precision: 0 })}{' '}
+
+              <div className="text-sm text-grey-500 dark:text-grey-300">
+                {formatWad(e.tokenCount, { precision: 0 })}{' '}
                 {tokenSymbolText({
                   tokenSymbol,
                   capitalize: false,
@@ -86,15 +88,8 @@ export default function ReservesEventElem({
             </div>
           ))}
 
-          {event.beneficiaryTicketAmount?.gt(0) && (
-            <div
-              style={{
-                display: 'flex',
-                justifyContent: 'space-between',
-                alignItems: 'baseline',
-              }}
-              className="text-sm"
-            >
+          {event.beneficiaryTokenCount?.gt(0) && (
+            <div className="flex items-baseline justify-between">
               <div>
                 <EthereumAddress
                   className="text-grey-900 dark:text-slate-100"
@@ -102,14 +97,9 @@ export default function ReservesEventElem({
                 />
                 :
               </div>
-              <div className="text-grey-500 dark:text-grey-300">
-                {formatWad(event.beneficiaryTicketAmount, {
+              <div className="text-sm text-grey-500 dark:text-grey-300">
+                {formatWad(event.beneficiaryTokenCount, {
                   precision: 0,
-                })}{' '}
-                {tokenSymbolText({
-                  tokenSymbol,
-                  capitalize: false,
-                  plural: true,
                 })}
               </div>
             </div>
