+++ src/components/v1/V1Project/ProjectActivity/index.tsx
@@ -1,12 +1,7 @@
 import { DownloadOutlined } from '@ant-design/icons'
 import { t, Trans } from '@lingui/macro'
 import { Button, Space } from 'antd'
-import AddToBalanceEventElem from 'components/activityEventElems/AddToBalanceEventElem'
-import BurnEventElem from 'components/activityEventElems/BurnEventElem'
-import DeployedERC20EventElem from 'components/activityEventElems/DeployedERC20EventElem'
-import PayEventElem from 'components/activityEventElems/PayEventElem'
-import ProjectCreateEventElem from 'components/activityEventElems/ProjectCreateEventElem'
-import RedeemEventElem from 'components/activityEventElems/RedeemEventElem'
+import { AnyProjectEvent } from 'components/activityEventElems'
 import { JuiceListbox } from 'components/inputs/JuiceListbox'
 import Loading from 'components/Loading'
 import SectionHeader from 'components/SectionHeader'
@@ -16,9 +11,6 @@ import { V1ProjectContext } from 'contexts/v1/Project/V1ProjectContext'
 import { ProjectEventFilter, useProjectEvents } from 'hooks/useProjectEvents'
 import { useContext, useMemo, useState } from 'react'
 
-import ReservesEventElem from './eventElems/ReservesEventElem'
-import TapEventElem from './eventElems/TapEventElem'
-import V1ConfigureEventElem from './eventElems/V1ConfigureEventElem'
 import { V1DownloadActivityModal } from './V1DownloadActivityModal'
 
 const PAGE_SIZE = 10
@@ -26,6 +18,7 @@ const PAGE_SIZE = 10
 export function V1ProjectActivity() {
   const { projectId } = useContext(ProjectMetadataContext)
   const { tokenSymbol } = useContext(V1ProjectContext)
+  const [isFetchingMore, setIsFetchingMore] = useState<boolean>()
 
   const [downloadModalVisible, setDownloadModalVisible] = useState<boolean>()
   const [eventFilter, setEventFilter] = useState<ProjectEventFilter>('all')
@@ -42,59 +35,27 @@ export function V1ProjectActivity() {
     first: PAGE_SIZE,
   })
 
+  const isLoading = loading || isFetchingMore
+
   const projectEvents = data?.projectEvents
 
   const count = projectEvents?.length || 0
 
   const list = useMemo(
     () =>
-      projectEvents?.map(e => {
-        let elem: JSX.Element | undefined = undefined
-
-        if (e.payEvent) {
-          elem = <PayEventElem event={e.payEvent} />
-        }
-        if (e.burnEvent) {
-          elem = <BurnEventElem event={e.burnEvent} tokenSymbol={tokenSymbol} />
-        }
-        if (e.addToBalanceEvent) {
-          elem = <AddToBalanceEventElem event={e.addToBalanceEvent} />
-        }
-        if (e.tapEvent) {
-          elem = <TapEventElem event={e.tapEvent} />
-        }
-        if (e.redeemEvent) {
-          elem = <RedeemEventElem event={e.redeemEvent} />
-        }
-        if (e.printReservesEvent) {
-          elem = <ReservesEventElem event={e.printReservesEvent} />
-        }
-        if (e.projectCreateEvent) {
-          elem = <ProjectCreateEventElem event={e.projectCreateEvent} />
-        }
-        if (e.deployedERC20Event) {
-          elem = <DeployedERC20EventElem event={e.deployedERC20Event} />
-        }
-        if (e.v1ConfigureEvent) {
-          elem = <V1ConfigureEventElem event={e.v1ConfigureEvent} />
-        }
-
-        if (!elem) return null
-
-        return (
-          <div
-            className="mb-5 border-b border-smoke-200 pb-5 dark:border-grey-600"
-            key={e.id}
-          >
-            {elem}
-          </div>
-        )
-      }),
+      projectEvents?.map(e => (
+        <div
+          className="mb-5 border-b border-smoke-200 pb-5 dark:border-grey-600"
+          key={e.id}
+        >
+          <AnyProjectEvent event={e} tokenSymbol={tokenSymbol} />
+        </div>
+      )),
     [projectEvents, tokenSymbol],
   )
 
   const listStatus = useMemo(() => {
-    if (loading) {
+    if (isLoading) {
       return (
         <div>
           <Loading />
@@ -102,7 +63,7 @@ export function V1ProjectActivity() {
       )
     }
 
-    if (count === 0 && !loading) {
+    if (count === 0 && !isLoading) {
       return (
         <div className="border-t border-smoke-200 pt-5 text-grey-500 dark:border-grey-600 dark:text-grey-300">
           <Trans>No activity yet</Trans>
@@ -114,13 +75,14 @@ export function V1ProjectActivity() {
       return (
         <div
           className="cursor-pointer text-center text-grey-500 dark:text-grey-300"
-          onClick={() =>
+          onClick={() => {
+            setIsFetchingMore(true)
             fetchMore({
               variables: {
                 skip: count,
               },
-            })
-          }
+            }).finally(() => setIsFetchingMore(false))
+          }}
         >
           <Trans>Load more</Trans>
         </div>
@@ -132,7 +94,7 @@ export function V1ProjectActivity() {
         <Trans>{count} total</Trans>
       </div>
     )
-  }, [loading, fetchMore, count])
+  }, [isLoading, fetchMore, count])
 
   return (
     <div>
