+++ src/lib/apollo/client.ts
@@ -39,8 +39,12 @@ const client = new ApolloClient({
       Query: {
         fields: {
           projectEvents: {
-            keyArgs: [],
+            // Create a new cache list when the `where` variable changes. Otherwise apply the below `merge` function to any existing cached list, even if other variables change like `first` or `skip`. This strategy enables infinite paging for unique lists while still preventing cached data from being unexpectedly included in query results, especially when updating the `where` argument to filter events by Paid, Redeemed, etc.
+            // https://www.apollographql.com/docs/react/pagination/key-args/
+            keyArgs: ['where'],
             merge(existing = [], incoming) {
+              // For a single list (even if `first` or `skip` arguments have changed) concatenate results of subsequent queries into the existing cache
+              // https://www.apollographql.com/docs/react/pagination/core-api
               return [...existing, ...incoming]
             },
           },
